* The new MyCollab edition with many updates:
- Use date time of java 8 (replace the vanilla Date api of java 7 and joda time)
- Upgrade to Vaadin 8 and change a lot in UI layer
- Mobile friendly. No two different versions for mobile and desktop
- Remove the CRM and File module
* Update libraries versions

+++ mycollab-web/src/main/java/com/mycollab/module/project/view/task/TaskPreviewForm.java
@@ -139,7 +139,7 @@ public class TaskPreviewForm extends AdvancedPreviewBeanForm<SimpleTask> {
                     @Subscribe
                     public void handle(TaskEvent.NewTaskAdded event) {
                         final ProjectTaskService projectTaskService = AppContextUtil.getSpringBean(ProjectTaskService.class);
-                        SimpleTask task = projectTaskService.findById((Integer) event.getData(), AppUI.getAccountId());
+                        SimpleTask task = projectTaskService.findById(event.getData(), AppUI.getAccountId());
                         if (task != null && tasksLayout != null) {
                             tasksLayout.addComponent(generateSubTaskContent(task), 0);
                         }
